script: Ensure scheduled timers account for suspension offset. (#39880)
When pages go into the bfcache, rendering engines record when the page
is suspended. As part of
https://html.spec.whatwg.org/multipage/browsing-the-web.html#reactivate-a-document,
any outstanding timers need to have their duration offset by the time
since the page was suspended. We used the correct offset when scheduling
new timers in a page that has been reactivated, but the wrong offset
when dealing with outstanding timers.

Testing: Lots of WPT tests that either pass or no longer timeout.
Fixes: #39545

Signed-off-by: Josh Matthews <josh@joshmatthews.net>