webdriver: Deserialize `WebWindow` and report error correctly for script execution (#39885)
We were relying on `fn WebdriverWindow` to report `NoSuchWindow` during
script execution, regardless of the fact that it is never `Fallible`. I
was really hoping to refactor this to return `Fallible` to reduce one
RTT of IPC, but there is no easy way to convert between DOM exception
https://github.com/servo/servo/blob/56b806b129bf7111e6442d3fc2bd08d602abb951/components/script_bindings/error.rs#L15
and JavaScriptEvaluationError
https://github.com/servo/servo/blob/56b806b129bf7111e6442d3fc2bd08d602abb951/components/shared/embedder/lib.rs#L1092.

We also migrate `element.rs` to `script_argument_extraction.rs` to
reflect the true purpose of the file.

Testing: 10 new subtests passing.

---------

Signed-off-by: Euclid Ye <yezhizhenjiakang@gmail.com>