layout: Let min-content size break before each unbreakable segment run (#39830)
We were only allowing a soft wrap opportunity before the first run in a
`TextRunSegment`, but not before the other runs. For example, this meant
that the min-content size of `中文中文` would be 4ic instead of just 1ic.

The change brings the logic in `ContentSizesComputation::process_item()`
closer to `TextRunSegment::layout_into_line_items()`.

Testing: Adding new tests. There is also a new failure because we don't
support `word-break: auto-phrase`.
Fixes: #39728

---------

Signed-off-by: Richard Tjokroutomo <richard.tjokro2@gmail.com>
Signed-off-by: Oriol Brufau <obrufau@igalia.com>
Co-authored-by: Oriol Brufau <obrufau@igalia.com>