layout: Add `Contents::Widget(NonReplacedContents)` for HTML widgets
Widgets like `<input>` and `<textarea>` have various behaviors similar
to replaced elements, but aren't fully replaced.

Therefore, this adds a new case into the `Contents` enum so that it can
represent the three states: non-replaced, fully replaced, and widget.
This means that `NonReplacedContents::OfTextControl` can be dropped.

The code is refactored in order to centralize the logic to identify
widgets in a single place.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>