layout: Use the new outset-adjusted border radius algorithm
When inflating the size of a rectangle with rounded corners, e.g. using
a spreading box shadow or `overflow-clip-margin`, we need to compute
the radii of the rounded corners of the resulting shape.

We were previously using a discontinuous algorithm, like Gecko and
WebKit. Some years ago, the spec came up with a continuous algorithm,
but the behavior wasn't ideal in some cases. More recently, the CSSWG
agreed on a new better algorithm, which is what this PR implements.

This way we align with Blink, which has already shipped this.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>