layout: Cleanup clear_fragment_layout_cache()
`LayoutBox::clear_fragment_layout_cache()` was just redirecting to the
`clear_fragment_layout_cache()` method of inner structures, until
reaching `LayoutBoxBase` where the work is actually done.

So this patch gets the `LayoutBoxBase` first, and then calls the method
on it. To do so, `LayoutBox::with_base_fold()` is added as a non-mutable
version of the existing `LayoutBox::with_base_mut_fold()`. But since
we the return value doesn't matter, `LayoutBox::with_each_base()` is
also added as a tiny helper around `LayoutBox::with_base_fold()`.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>