servoview: Focus the view when a touch input occurs. (#40049)
Focusing the location bar in the Android app shows the onscreen
keyboard, but clicking on a link in the current document does not make
it disappear. This change ensures the input focus is reset appropriately
and ensures the keyboard disappear when the location bar is not being
edited.

Testing: Manually tested. No automated tests for Android app at this
time.
Fixes: #40008

Signed-off-by: Josh Matthews <josh@joshmatthews.net>