compositor: Have viewport meta tag affect the layout viewport (#40055)
The CSS specifications defines two viewports[^1], the "layout viewport"
and the visual viewport. The layout viewport affects layout and the
visual viewport is the result of applying pinch zoom. This changes makes
it so that the initial-scale property of `<meta type="viewport">`
affects the layout viewport.

Testing: Added a new unit test

[^1]: https://www.w3.org/TR/css-viewport-1/#the-viewport

Signed-off-by: atbrakhi <atbrakhi@igalia.com>
Co-authored-by: Shubham Gupta <shubham13297@gmail.com>
Co-authored-by: Martin Robinson <mrobinson@igalia.com>