xpath: Provide error message when throwing SyntaxError from parsing (#40103)
This change was split of from https://github.com/servo/servo/pull/39977
to make that PR easier to review.

The `Debug` impl for the error is not very useful yet, but
https://github.com/servo/servo/pull/39977 swaps out the error type
itself, so there's no point in improving it here.

Testing: I've manually tested that the error message appears. I don't
think we have tests for error messages.
Part of #34527

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>