libservo: Stop using the custom test harness and `run_api_tests`
The changes in #39812 seem to have completely disabled the API tests, so
this PR gets them running again by removing the custom test harness
(apart from the multiprocess test). The custom test harness is typically
not necessary now that all tests are run in their own process.

In addition, the behavior of the `WebView::page_zoom()` is fixed when
set via the `viewport` meta tag. This functionality was broken, but the
unit test wasn't catching it.

There is still one failing test, but that is tracked via #40129.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>