tracing: Don't trace maybe_close_random_pipeline
Most of the time this function does nothing
and returns immediately,
so there isn't really a point in tracing it.
If one wanted to know when a pipeline is closed,
one could add a trace point **after** the early return.
However during review it was suggested that we currently don't need this
information and thus just remove the trace.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>