script: Replace IndexedDB's unsafe uses of RootedVec<JSVal>. (#40143)
While #40141 is filed to address the complete unsafe pattern, fixing the
instances of this within the IndexedDB code is something that can happen
today. It is not safe to store raw GC values inside of RootedVec; they
need to be inside Heap wrappers, and those can't be stored in a vector
that can reallocate without being wrapped inside of a Box.

Testing: Fixes an intermittent crashing test.
Fixes: #39766

Signed-off-by: Josh Matthews <josh@joshmatthews.net>