compositing: Initialize the WebRender image handlers with the IOCompositor (#40134)
In order to move all WebRender initialization closer to the renderer,
so that it can support multiple WebRender instances, move all external
image handler initialiazation into `compositing`.

This also modifies the order of initalization. Now the `IOCompositor`
(renderer) starts before the `Constellation`. This will be useful
when Servo construction can fail, so that we fail before starting a
`Constellation` thread.

Finally, replace all occurences of "Webrender" with "WebRender".

Testing: This should not change behavior so is covered by existing
tests.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Mukilan Thiyagarajan <mukilan@igalia.com>