xpath: Throw a `NAMESPACE_ERR` DOMException when namespace resolver callback returns invalid value or throws (#40167)
I always thought that exceptions thrown in namespace resolver callbacks
should be propagated to the call site of `document.evaluate`.
That is not the case - the exception is silently swallowed and you get a
namespace error, the same as if the callback had returned any other
invalid value.

Testing: New web platform tests start to pass
Fixes https://github.com/servo/servo/issues/39931
Part of #34527

---------

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>