libservo: Expose `WebViewPoint` / `WebViewRect` and use them for the API
This change exposes two new data types in the API, which allow embedders
to supply coordinates in either page pixels or device pixels. In
particular, WebDriver wants to give coordiantes to the API in page
pixels. This prevents the embedder from having to do these conversions
and makes it harder to misuse the API.

Co-authored-by: Rakhi Sharma <atbrakhi@igalia.com>
Signed-off-by: Martin Robinson <mrobinson@igalia.com>