script: Use new inserted node when performing slot checks (#40251)
https://github.com/servo/servo/blob/main/components/script/dom/node.rs#L2560
The node in this line should be kid.
Step 7.4 If parent is a shadow host whose shadow root’s slot assignment
is "named" and node is a slottable, then assign a slot for node.
The node here should be the node of "for each" rather than the node of
the insert parameter.

Fixes: https://github.com/servo/servo/issues/40250