libservo: Expose `WebViewPoint` / `WebViewRect` and use them for the API (#40223)
This change exposes two new data types in the API, which allow embedders
to supply coordinates in either page pixels or device pixels. In
particular, WebDriver wants to give coordiantes to the API in page
pixels. This prevents the embedder from having to do these conversions
and makes it harder to misuse the API.

Testing: This should fix WebDriver conformance tests that are run in
headed mode.
Fixes: #40152

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Rakhi Sharma <atbrakhi@igalia.com>