libservo: Use the embedder control API for simple dialogs (#40239)
This is part of a series of changes to unify the different APIs we have
currently for requesting the embedder to display UI elements.

This patch doesn't change how the script requests the dialogs to be
displayed, it only switches the webview delegate method that gets
invoked in response to the request from script. The change also adds an
`EmbedderControlId` field to the simple dialogs, but these are not
currently used. They exist only to satify the calls to `id()` method of
the `EmbedderControl` enum.

Testing: Added new unit tests to validate that the dialogs are
displayed.

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>