wpt: Enable WebDriver by default when running WPT
This change switches the default test runner for WPT to be WebDriver,
enabling testdriver tests by default. In addition, it update results to
reflect the ones that you would expect when running with WebDriver.
While there are some failures that require more investigation, in
general the differences in results are fairly explicable.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>