wpt: Make legacy `ServoWdspecBrowser` work as intended (#40296)
https://github.com/servo/servo/pull/40289#discussion_r2479889037 intend
to process `headless` param properly.

For either `servo` or `servodriver` binary, the expected behaviour is:
- When given one test, it should depends on cmd arg `--headless`
- With multiple tests, we always default to `--headless`

The subtle point is: `ServoWdspecBrowser` didn't process `headless`
param at all and always default to headless. Also, `executor_kwargs`
does not work for it but only `browser_kwargs` works as it does not use
`ServoExecutor`.

Testing: Manually tested with legacy driver, running in single/multiple
tests, with/without headless param, with wdspec or none-wdspec test.
Fixes: https://github.com/servo/servo/pull/40289#discussion_r2479889037

---------

Signed-off-by: Euclid Ye <yezhizhenjiakang@gmail.com>