script: Trigger rendering update for skipped resize observations. (#40353)
When resize observations are skipped, they get delayed until the next
turn of the event loop. In headless tests, we're not guaranteed that
this next turn occurs in time, unless some other event happens to run.
This leads to intermittent results unless we ensure that the event loop
will run again.

Testing: Newly passing test.
Fixes: #33599

Signed-off-by: Josh Matthews <josh@joshmatthews.net>