wpt: Filter non-stable expected results and create an artifact for it
This change ensures that the filtered raw log file of WPT results only
contains stable unexpected results and uploads a resulting aggregated
log file as an artifact. This artifact can be used to generate new
baselines of *only* stable unexpected results, so is useful for updating
results given a CI run. A future change will add a `mach` command to do
this automatically.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>