tools: Remove wrapper around message in the JSON output of the Devtools parser
Instead of outputing `{"_to": "...", "message": {"to": "...", ...}}` the
JSON message is simplified to `{"to": "...", ...}`. This is possible
since all valid devtools messages include a "to" or a "from" field. These
two fields will always be placed at the start of the serialized message,
with the rest of fields being sorted alphabetically.

Simplify detection of server or client message, removing the need to pass
the devtools port when reading a .pcap file. Now the pretty message says
"Server" and "Client" instead of "Servo" and "Firefox", which was
innacurate when inspecting one instance of Firefox from another.

Signed-off-by: eri <eri@igalia.com>