compositing: Allow mapping from WebRender key types to `PainterId` (#40484)
In order to properly select the appropriate `Painter` when dealing with
incoming keys to the renderer API, we ensure that these keys carry the
`PainterId` encoded as `IdNamespace` inside them. This should eliminate
the need to pass the `WebViewId` or `PainterId` with messages involving
these keys.

Testing: This should not change observable behavior, so should be
covered
by existing tests.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Mukilan Thiyagarajan <mukilan@igalia.com>