storage: Introduce a new storage coordination thread
This patch introduces a new thread responsible for coordinating all current
and future storage endpoints.

The prefix "client" was chosen to make it explicit that this functionality
applies to client side, locally persisted storage, and should not be confused
with the localStorage API or the StorageManager interface.

Testing: Unit tests continue to pass.

Signed-off-by: Jan Varga <jvarga@igalia.com>