dom: Add the attribute mutation reason
For some specific cases (one of them see below) where are
requirement to identify the reason of the attribute mutation
(by whom/why it was changed - by cloning, parser or directly),
so the following `reason` parameter was added to
`AttributeMutation::Set` option.

Note that for the HTMLMediaElement the internal `muted` state
should be set to true when the element is created and the
element has a `muted` content attribute specified
(should be done once on the first attribute mutation caused
by parser or cloning).

See https://html.spec.whatwg.org/multipage/#dom-media-muted

Testing: Improvements in the following tests
- html/semantics/embedded-content/media-elements/user-interface/muted.html

Signed-off-by: Andrei Volykhin <andrei.volykhin@gmail.com>