ohos: Make test-speedometer-ohos more robust
The main change is running `hdc shell power-shell setmode 602`,
which sets the device to performance mode, increasing the screen timeout
and manually waking the device up.
We add a helper class for that, so that we can use the `with` syntax
to undo these changes at the end of scope.
We also add a small HDC wrapper class, which slightly simplifies
and abstracts over some common functionality like running commands
or making a screenshot.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>